v00r03: manifest.bml exporting with icarus and daedalus
Imported daedalus, a version of icarus that generates manifests for
higan v095.

Mercurial Magic now has the option of creating a manifest when exporting
a Game Pak, but only if icarus is present in the same directory as
Mercurial Magic. If daedalus is also present, Mercurial Magic will
create a hybrid manifest that can work on both higan v095 and higan
v096-v099 and up.