v01r01: hiro GTK/Qt compilation error fix
Fixed a long-standing bug in hiro preventing Mercurial Magic from
compiling on Linux or BSD. Credit to SuperMikeMan for the fix.

Mercurial Magic's version of hiro is the same as in nSide, which
includes a fix for the Keyboard fetching the wrong keys. Strictly
speaking, it is not necessary for Mercurial Magic to have this version
of hiro, but the fact remains that higan's version of hiro has the wrong
keys bug, which is why Mercurial Magic has nSide's version.