v01r05: bpspatch file lock bugfix, clear paths after finish
Replaced the bpspatch object with a unique_pointer to the bpspatch
object so that it can be reinitialized each time a new ROM is selected
and when exporting finishes successfully.

Also made the finishExport function clear both text boxes after a
successful export operation. If an export operation fails, the text
boxes will retain their contents, and the BPS patch will remain in
memory.