v01r09: SD2SNES/Snes9x Force Manifest Creation option
Added an option to force the creation of a manifest even when using the
SD2SNES/Snes9x export method. This option is meant for quick testing of
every supported target (higan, Snes9x, SD2SNES) using only a single
cartridge folder. As such, it will only be available if Mercurial Magic
can find both icarus and daedalus; just one or the other will not enable
the option by itself.