v01r10: Prohibit multi-patching
On byuu's forum (https://board.byuu.org/viewtopic.php?f=8&t=1829&start=40#p47798), it has been determined that BPS cannot reliably support multi-patching when a patch increases a ROM's size but another expects that ROM to be its original size. As such, ignoring the CRC32 is now a violation of the BPS spec.

Mercurial Magic will now terminate an export if a ROM's CRC32 does not match the patch's.